+++ src/Service/S3/src/ValueObject/EventBridgeConfiguration.php
@@ -2,15 +2,13 @@
 
 namespace AsyncAws\S3\ValueObject;
 
-use AsyncAws\S3\ValueObject\EventBridgeConfiguration as EventBridgeConfiguration1;
-
 /**
  * A container for specifying the configuration for Amazon EventBridge.
  */
 final class EventBridgeConfiguration
 {
     /**
-     * @param array|EventBridgeConfiguration1 $input
+     * @param array|EventBridgeConfiguration $input
      */
     public static function create($input): self
     {